e commerce/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.48
pymysql==1.1.1
pydantic[email]==2.7.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
python-dotenv==1.0.1
razorpay==1.4.1
Pillow==10.3.0
